Implement decom_all method
